Submit transactions to follower
We implement this functionality using a polling leader. Every follower
(including the leader) stores new transactions in a temporary buffer.
This buffer is emptied every time a polling request arrives and the
content of the buffer is sent to the leader. Followers are responsible
for checking whether they are sending the real leader. Otherwise a
malicious polling request may cause the follower to empty its buffer
prematurely.